requirements.txt
pytest
ruff
pydantic
pyyaml
httpx
pytest-timeout
websocket-client
